% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteCode.R
\docType{data}
\name{MetroPalettes}
\alias{MetroPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 13.
}
\usage{
MetroPalettes
}
\description{
Use names(MetPalettes) to return all possible palette names. Current choices are:
\code{Paris}, and \code{Madrid}.
Use \code{\link{metro.brewer}} to construct palettes.
}
\keyword{datasets}
