% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteCode.R
\name{metro.brewer}
\alias{metro.brewer}
\title{Metro Palette Generator}
\usage{
metro.brewer(
  palette_name,
  n,
  direction = c(1, -1),
  override_order = FALSE,
  return_hex = FALSE
)
}
\arguments{
\item{palette_name}{Name of Palette. Choices are:}

\item{n}{Number of desired colors. If number of requested colors is beyond the scope of the palette,
colors are automatically interpolated. If n is not provided, the length of the palette is used.}

\item{direction}{Sets order of colors. Default palette is 1. If direction is -1, palette color order is reversed}

\item{override_order}{Colors are picked from palette to maximize readability and aesthetics. This means
that colors are not always selected in sequential order from the full palette. If override_order is set to TRUE,
colors are selected in sequential order from the full palette instead. Default is FALSE.}

\item{return_hex}{Default is FALSE. If TRUE, hex codes of colors will be returned in addition to the palette.}
}
\value{
A vector of colors.
}
\description{
Color palettes inspired by subway and train maps. Complete palette colors and
the maps that inspired them can be found \href{https://github.com/juliagdealedo/MetroBrewer}{on Github}.
}
\examples{
metro.brewer("Madrid")

metro.brewer("London", direction=-1)

metro.brewer("Sydney", 4, override_order=TRUE)


}
\keyword{colors}
