% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteCode.R
\docType{data}
\name{MetroPalettes}
\alias{MetroPalettes}
\title{Complete list of palettes.}
\format{
An object of class \code{list} of length 17.
}
\usage{
MetroPalettes
}
\description{
Use \code{names(MetroPalettes)} to return all possible palette names.
Use \code{\link{metro.brewer}} to construct palettes.
}
\keyword{datasets}
